+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Added `spicyweb\oddsandends\fields\DisabledUsers` - Users (Disabled) field type
+
 ## 4.2.0 - 2023-03-27
 
 ### Added